+++ packages/plugin-browser/src/browser.ts
@@ -1,0 +1,8 @@
+import { createPlugin } from "@flint.fyi/core";
+
+import alerts from "./rules/alerts.js";
+
+export const browser = createPlugin({
+	name: "browser",
+	rules: [alerts],
+});