+++ packages/plugin-jsx/src/jsx.ts
@@ -1,0 +1,8 @@
+import { createPlugin } from "@flint.fyi/core";
+
+import accessKeys from "./rules/accessKeys.js";
+
+export const jsx = createPlugin({
+	name: "jsx",
+	rules: [accessKeys],
+});