+++ packages/ts/src/index.ts
@@ -1,3 +1,4 @@
 export { getTSNodeRange } from "./getTSNodeRange.js";
 export * from "./language.js";
 export { ts } from "./plugin.js";
+export { isGlobalDeclaration } from "./utils/isGlobalDeclaration.js";