+++ packages/ts/src/utils/isGlobalDeclaration.ts
@@ -1,0 +1,12 @@
+import * as ts from "typescript";
+
+import { declarationsIncludeGlobal } from "./declarationsIncludeGlobal.js";
+
+export function isGlobalDeclaration(
+	node: ts.Expression,
+	typeChecker: ts.TypeChecker,
+): boolean {
+	const declarations = typeChecker.getSymbolAtLocation(node)?.getDeclarations();
+
+	return !!declarations && declarationsIncludeGlobal(declarations);
+}